refactor(ingest): move the staging cluster out of tasks_common.py (#2026)
tasks_common.py was the largest capped backend module at 2423 lines. The
staging acquisition and cleanup half now lives in tasks_staging.py: the
StagingResult bundle, the two staging-object reapers, the upload-safety
gauntlet, the original-file archive, the post-ogr2ogr staging pipeline, and
the shared terminal-failure handler.

Pure move. Every moved definition is byte-identical under ast.dump(), and so
is every definition left behind. No re-export shim: a patch aimed at
tasks_common would have kept passing while silently missing the call, so each
of the string-path patch targets moved to the module that now binds the name.

Closes #2026
