+++ backend/tests/finding_markers.py
@@ -356,13 +356,14 @@ UNANCHORED_MARKER_DEBT: dict[str, int] = {
     "processing/ingest/presigned.py": 1,
     "processing/ingest/router.py": 11,
     "processing/ingest/schemas.py": 1,
-    "processing/ingest/tasks_common.py": 42,
+    "processing/ingest/tasks_common.py": 36,
     "processing/ingest/tasks_postgis_refresh.py": 2,
     "processing/ingest/tasks_raster.py": 35,
     "processing/ingest/tasks_raster_common.py": 6,
     "processing/ingest/tasks_raster_replace.py": 4,
     "processing/ingest/tasks_reupload.py": 7,
     "processing/ingest/tasks_stac_refresh.py": 4,
+    "processing/ingest/tasks_staging.py": 6,
     "processing/ingest/tasks_vector.py": 32,
     "processing/ingest/tasks_vrt.py": 7,
     "processing/ingest/url_fetch.py": 1,
