+++ backend/tests/test_ingest.py
@@ -1137,7 +1137,7 @@ async def test_archive_original_file_success_leaves_user_metadata_untouched(
     mock_storage.put = AsyncMock(side_effect=lambda key, fobj: put_calls.append(key))
 
     monkeypatch.setattr(
-        "app.processing.ingest.tasks_common.get_storage", lambda: mock_storage
+        "app.processing.ingest.tasks_staging.get_storage", lambda: mock_storage
     )
 
     mock_session = AsyncMock()
@@ -1179,7 +1179,7 @@ async def test_archive_original_file_failure_marks_user_metadata(tmp_path, monke
     mock_storage.put = AsyncMock(side_effect=RuntimeError("S3 unreachable"))
 
     monkeypatch.setattr(
-        "app.processing.ingest.tasks_common.get_storage", lambda: mock_storage
+        "app.processing.ingest.tasks_staging.get_storage", lambda: mock_storage
     )
 
     mock_session = AsyncMock()
@@ -1230,7 +1230,7 @@ async def test_archive_original_file_commit_failure_does_not_raise(
     mock_storage = AsyncMock()
     mock_storage.put = AsyncMock(side_effect=RuntimeError("S3 unreachable"))
     monkeypatch.setattr(
-        "app.processing.ingest.tasks_common.get_storage", lambda: mock_storage
+        "app.processing.ingest.tasks_staging.get_storage", lambda: mock_storage
     )
 
     mock_session = AsyncMock()
@@ -1276,7 +1276,7 @@ async def test_archive_original_file_failure_truncates_long_error(
     mock_storage.put = AsyncMock(side_effect=RuntimeError(long_error))
 
     monkeypatch.setattr(
-        "app.processing.ingest.tasks_common.get_storage", lambda: mock_storage
+        "app.processing.ingest.tasks_staging.get_storage", lambda: mock_storage
     )
 
     mock_session = AsyncMock()
