+++ backend/tests/test_staging_pipeline.py
@@ -868,7 +868,7 @@ class TestIngestVectorIntoStagingErrors:
                 return_value=[],
             ),
             patch(
-                "app.processing.ingest.tasks_common._detect_and_override_geometry",
+                "app.processing.ingest.tasks_staging._detect_and_override_geometry",
                 new_callable=AsyncMock,
                 side_effect=RuntimeError("geometry override failed"),
             ),
